Rename jmxmetricextension to jmxmetricreceiver
