+++ Makefile
@@ -17,8 +17,8 @@ OTEL_VERSION=master
 # Modules to run integration tests on.
 # XXX: Find a way to automatically populate this. Too slow to run across all modules when there are just a few.
 INTEGRATION_TEST_MODULES := \
-	extension/jmxmetricextension/ \
 	receiver/dockerstatsreceiver \
+	receiver/jmxmetricreceiver/ \
 	receiver/redisreceiver \
 	internal/common
